+++ lib/dusk.dart
@@ -15,6 +15,7 @@
 library;
 
 export 'src/dusk_artisan_provider.dart';
+export 'src/dusk_error_capture.dart' show recentCapturedExceptions;
 export 'src/dusk_navigate_adapter.dart';
 export 'src/dusk_plugin.dart';
 export 'src/dusk_snapshot_enricher.dart';